Create Orchestrator class
